requirements.txt
fastapi
uvicorn
pytest
pytest-anyio
httpx
httpx-ws